refactor(cli): move directoryUtils.ts to common folder
