+++ packages/cli/src/common/utils/directoryUtils.ts
